feat: move config list filter control from toolbar
