+++ src/pages/config/ConfigList.tsx
@@ -1,5 +1,4 @@
-import React, { useEffect, useState, useMemo } from "react";
-import { debounce } from "lodash";
+import { useCallback, useEffect } from "react";
 import {
   useNavigate,
   useSearchParams,
@@ -8,23 +7,14 @@ import {
 import { getAllConfigs } from "../../api/services/configs";
 import { filterConfigsByText } from "../../components/ConfigViewer/utils";
 import { BreadcrumbNav } from "../../components/BreadcrumbNav";
-import { TextInputClearable } from "../../components/TextInputClearable";
 import ConfigList from "../../components/ConfigList";
-import { SearchSelectTag } from "../../components/SearchSelectTag";
-import { QueryBuilder } from "../../components/QueryBuilder";
-import { Switch } from "../../components/Switch";
-import { RefreshButton } from "../../components/RefreshButton";
 import { useLoader } from "../../hooks";
 import { useConfigPageContext } from "../../context/ConfigPageContext";
-import { ReactSelectDropdown } from "../../components/ReactSelectDropdown";
-
-const ConfigFilterViewTypes = {
-  basic: "Basic",
-  advanced: "Advanced"
-};
+import ConfigsListFilters from "../../components/ConfigsListFilters";
+import { RefreshButton } from "../../components/RefreshButton";
 
 export function ConfigListPage() {
-  const [params, setParams] = useSearchParams();
+  const [params] = useSearchParams();
   const navigate = useNavigate();
   const {
     configState: { data, filteredData },
@@ -32,34 +22,18 @@ export function ConfigListPage() {
   } = useConfigPageContext();
   const { loading, setLoading } = useLoader();
   const { setTitle, setTitleExtras } = useOutletContext<any>();
-  const [configFilterView, setConfigFilterView] = useState(
-    params.get("query")
-      ? ConfigFilterViewTypes.advanced
-      : ConfigFilterViewTypes.basic
-  );
-  const options = useMemo(() => {
-    return [ConfigFilterViewTypes.basic, ConfigFilterViewTypes.advanced];
-  }, []);
 
-  const configTagItems: any = useMemo(() => {
-    if (!data) return [];
-    data.flatMap((d) => {
-      return Object.entries(d?.tags || {});
-    });
-  }, [data]);
+  useEffect(() => {
+    setTitleExtras(
+      <RefreshButton onClick={() => getAllConfigs()} animate={loading} />
+    );
+  }, [loading, setTitleExtras]);
 
   const search = params.get("search");
   const tag = decodeURIComponent(params.get("tag") || "All");
   const configType = decodeURIComponent(params.get("type") || "All");
 
-  useEffect(() => {
-    if (params.get("query")) {
-      return;
-    }
-    fetchAllConfigs();
-  }, [configFilterView, params]);
-
-  function fetchAllConfigs() {
+  const fetchAllConfigs = useCallback(() => {
     setLoading(true);
     getAllConfigs()
       .then((res) => {
@@ -75,7 +49,14 @@ export function ConfigListPage() {
         console.error(err);
         setLoading(false);
       });
-  }
+  }, []);
+
+  useEffect(() => {
+    if (params.get("query")) {
+      return;
+    }
+    fetchAllConfigs();
+  }, [params]);
 
   const handleRowClick = (row?: { original?: { id: string } }) => {
     const id = row?.original?.id;
@@ -83,84 +64,6 @@ export function ConfigListPage() {
       navigate(`/configs/${id}`);
     }
   };
-
-  const extra = (
-    <div className="flex space-x-2 mr-4">
-      {configFilterView === ConfigFilterViewTypes.advanced && (
-        /* @ts-expect-error */
-        <QueryBuilder
-          refreshConfigs={(e: any) => {
-            setConfigState((state) => {
-              return {
-                ...state,
-                data: e
-              };
-            });
-          }}
-        />
-      )}
-
-      {configFilterView === ConfigFilterViewTypes.basic && (
-        <>
-          <RefreshButton onClick={() => fetchAllConfigs()} animate={loading} />
-          <TypeDropdown
-            value={configType}
-            onChange={(ct: any) => {
-              params.set("type", encodeURIComponent(ct || ""));
-              setParams(params);
-            }}
-          />
-
-          <span className="w-80">
-            <SearchSelectTag
-              tags={configTagItems}
-              value={tag}
-              onChange={(ct) => {
-                params.set("tag", encodeURIComponent(ct.value || ""));
-                setParams(params);
-              }}
-            />
-          </span>
-
-          {/* @ts-expect-error */}
-          <TextInputClearable
-            onChange={debounce((e) => {
-              const query = e.target.value || "";
-              params.set("search", query);
-              setParams(params);
-            }, 200)}
-            className="w-80"
-            placeholder="Search for configs"
-            defaultValue={params.get("search")}
-          />
-        </>
-      )}
-
-      {/* @ts-expect-error */}
-      <Switch
-        onChange={(e: string) => {
-          setConfigFilterView(e);
-          setParams({});
-        }}
-        options={options}
-        value={configFilterView}
-      />
-    </div>
-  );
-
-  useEffect(() => {
-    setTitleExtras(extra);
-    return () => setTitleExtras(null);
-  }, [
-    data,
-    configType,
-    tag,
-    search,
-    configTagItems,
-    options,
-    configFilterView,
-    loading
-  ]);
 
   useEffect(() => {
     let filteredData = data;
@@ -197,51 +100,15 @@ export function ConfigListPage() {
   }, [data, search, configType, tag]);
 
   return (
-    <ConfigList
-      data={filteredData!}
-      handleRowClick={handleRowClick}
-      isLoading={loading}
-    />
+    <div className="flex flex-col h-full space-y-6 px-4 py-4">
+      <div className="flex flex-row items-center">
+        <ConfigsListFilters loading={loading} />
+      </div>
+      <ConfigList
+        data={filteredData!}
+        handleRowClick={handleRowClick}
+        isLoading={loading}
+      />
+    </div>
   );
 }
-
-const TypeDropdown = ({ ...rest }) => {
-  const items = {
-    All: {
-      id: "All",
-      name: "All",
-      description: "All",
-      value: "All"
-    },
-    EC2Instance: {
-      id: "EC2Instance",
-      name: "EC2 Instance",
-      description: "EC2 Instance",
-      value: "EC2Instance"
-    },
-    Subnet: {
-      id: "Subnet",
-      name: "Subnet",
-      description: "Subnet",
-      value: "Subnet"
-    }
-  };
-
-  const [selected, setSelected] = useState<any>(Object.values(items)[0].value);
-
-  return (
-    <ReactSelectDropdown
-      items={items}
-      name="type"
-      onChange={(value) => setSelected(value)}
-      value={selected}
-      className="w-64"
-      prefix={
-        <div className="text-xs text-gray-500 mr-2 whitespace-nowrap">
-          Type:
-        </div>
-      }
-      {...rest}
-    />
-  );
-};
